Changed Images

+++ src/hudDisplay/FollowTextItem.java
@@ -16,7 +16,7 @@ public class FollowTextItem extends TextItem{
 	
 	public void render(Graphics2D g) {
 		if(isHUDactive())
-		getMyScoreFont().drawString(g, String.valueOf(getMyStat().getMyValue()),(int) spriteToFollow.getX() + getMyX(), (int) spriteToFollow.getY() + getMyY());
+		getMyScoreFont().drawString(g, String.valueOf((int) getMyStat().getMyValue()),(int) spriteToFollow.getX() + getMyX(), (int) spriteToFollow.getY() + getMyY());
 	}
 
 }
